+++ flutter_inappwebview/lib/src/in_app_webview/in_app_webview_controller.dart
@@ -1,11 +1,13 @@
 import 'dart:core';
 
 import 'package:flutter/services.dart';
+
 import 'package:flutter_inappwebview_platform_interface/flutter_inappwebview_platform_interface.dart';
 
 import '../print_job/main.dart';
 import '../web_message/main.dart';
 import '../web_storage/web_storage.dart';
+
 import 'android/in_app_webview_controller.dart';
 import 'apple/in_app_webview_controller.dart';
 
@@ -293,6 +295,15 @@ class InAppWebViewController {
 
   ///{@macro flutter_inappwebview_platform_interface.PlatformInAppWebViewController.clearFocus}
   Future<void> clearFocus() => platform.clearFocus();
+
+  ///{@macro flutter_inappwebview_platform_interface.PlatformInAppWebViewController.setInputMethodEnabled}
+  Future<void> setInputMethodEnabled(bool enabled) => platform.setInputMethodEnabled(enabled);
+
+  ///{@macro flutter_inappwebview_platform_interface.PlatformInAppWebViewController.showInputMethod}
+  Future<void> showInputMethod() => platform.showInputMethod();
+
+  ///{@macro flutter_inappwebview_platform_interface.PlatformInAppWebViewController.hideInputMethod}
+  Future<void> hideInputMethod() => platform.hideInputMethod();
 
   ///{@macro flutter_inappwebview_platform_interface.PlatformInAppWebViewController.setContextMenu}
   Future<void> setContextMenu(ContextMenu? contextMenu) =>