+++ flutter_inappwebview_ios/ios/Classes/InAppWebView/WebViewChannelDelegate.swift
@@ -677,6 +677,20 @@ public class WebViewChannelDelegate: ChannelDelegate {
             } else {
                 result(false)
             }
+            break
+        case .setInputMethodEnabled:
+            if let webView = webView {
+                let enabled = arguments!["enabled"] as! Bool
+                webView.setInputMethodEnabled(enabled: enabled)
+            }
+            result(true)
+            break
+        case .hideInputMethod:
+            if let webView = webView {
+                webView.hideInputMethod()
+            }
+            result(true)
+            break
         }
     }